+++ CHANGELOG.md
@@ -7,10 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- Draft message for each chat.
 - Display of group descriptions.
 - Reception of typing indicators.
 
 ### Changed
+- The sidebar now highlights the active chat.
 - Grab focus of the entry when replying to a message.
 - UI improvements to the channel information dialog, which now also displays phone number, disappearing messages timer and description.
 - Fix the messages being sent on shift+enter instead of ctrl+enter.